requirements.txt
numpy
opencv-python
matplotlib
torch
pillow
fire
facenet-pytorch
deep-sort-realtime